Fix the backwash issue by eliminating the bottom virtual site.
The bottom virtual site was causing "not full" nodes to be considered
full. Therefore, by eliminating it and lazily checking each time a
bottom row object was involved in a union, we would check if it is
now connected to the top virtual site. If it is, system has percolated.